Refactor reusable auth page widgets into new files
Refactor reusable legacy auth page widgets into new files
